v1.0.0 - optimization of code and logic
